clean-up: extract `should_implement_trait`
Also put `SelfKind` into `lib.rs`, since it's shared with
`wrong_self_convention` -- not `utils.rs`, because it's not really a
utility function
